+++ plugins/kiali-backend/src/clients/fetch.ts
@@ -123,11 +123,15 @@ export class KialiFetcher {
     return null;
   }
 
-  handleUnsuccessfulResponse(res: AxiosError): KialiFetchError {
+  handleUnsuccessfulResponse(
+    res: AxiosError,
+    endpoint?: string,
+  ): KialiFetchError {
     const message = res.message;
     const codeMessage = res.code ? `status (${res.code}) ` : '';
-    const url = res.config?.url || '';
+    const url = endpoint || res.config?.url || '';
     const urlMessage = url ? `when fetching "${url}" in "Kiali";` : '';
+    this.logger.warn(`Error response axios: ${JSON.stringify(res)}`);
     this.logger.warn(
       `Received ${res.status} ${codeMessage} ${urlMessage} body=[${message}]`,
     );