+++ plugins/kiali-backend/src/service/router.ts
@@ -1,21 +1,25 @@
 import { errorHandler } from '@backstage/backend-common';
+import { CatalogApi } from '@backstage/catalog-client';
+import {
+  CompoundEntityRef,
+  parseEntityRef,
+  stringifyEntityRef,
+} from '@backstage/catalog-model';
 import { Config } from '@backstage/config';
+import { InputError } from '@backstage/errors';
+import { getBearerTokenFromAuthorizationHeader } from '@backstage/plugin-auth-node';
 
-import express from 'express';
+import express, { Request } from 'express';
 import { Logger } from 'winston';
 
-import {
-  KUBERNETES_ANNOTATION,
-  KUBERNETES_LABEL_SELECTOR,
-  KUBERNETES_NAMESPACE,
-  readKialiConfigs,
-} from '@janus-idp/backstage-plugin-kiali-common';
+import { readKialiConfigs } from '@janus-idp/backstage-plugin-kiali-common';
 
 import { KialiApiImpl } from '../clients/KialiAPIConnector';
 
 export interface RouterOptions {
   logger: Logger;
   config: Config;
+  catalogApi: CatalogApi;
 }
 
 export type OverviewQuery = {
@@ -24,41 +28,71 @@ export type OverviewQuery = {
   overviewType?: string;
   duration?: number;
   direction?: string;
-  annotation?: { [key: string]: string };
 };
 
 export const makeRouter = (
   logger: Logger,
   kialiAPI: KialiApiImpl,
+  catalogApi: CatalogApi,
 ): express.Router => {
-  const getAnnotations = (query: any): { [key: string]: string } => {
-    const annotation: { [key: string]: string } = {};
-    annotation[KUBERNETES_ANNOTATION] =
-      query[encodeURIComponent(KUBERNETES_ANNOTATION)];
-    annotation[KUBERNETES_LABEL_SELECTOR] =
-      query[encodeURIComponent(KUBERNETES_LABEL_SELECTOR)];
-    annotation[KUBERNETES_NAMESPACE] =
-      query[encodeURIComponent(KUBERNETES_NAMESPACE)];
-    return annotation;
+  const getEntityByReq = async (req: Request<any>) => {
+    const rawEntityRef = req.body.entityRef;
+    if (rawEntityRef && typeof rawEntityRef !== 'string') {
+      throw new InputError(`entity query must be a string`);
+    } else if (!rawEntityRef) {
+      throw new InputError('entity is a required field');
+    }
+    let entityRef: CompoundEntityRef | undefined = undefined;
+
+    try {
+      entityRef = parseEntityRef(rawEntityRef);
+    } catch (error) {
+      throw new InputError(`Invalid entity ref, ${error}`);
+    }
+
+    logger.info(`entityRef: ${JSON.stringify(entityRef)}`);
+
+    const token =
+      getBearerTokenFromAuthorizationHeader(req.headers.authorization) ||
+      (req.cookies?.token as string | undefined);
+
+    logger.info(`Token: ${JSON.stringify(token)}`);
+    const entity = await catalogApi.getEntityByRef(entityRef, {
+      token: token,
+    });
+
+    if (!entity) {
+      throw new InputError(
+        `Entity ref missing, ${stringifyEntityRef(entityRef)}`,
+      );
+    }
+    return entity;
   };
+
   const router = express.Router();
   router.use(express.json());
 
-  router.get('/config', async (_, res) => {
+  router.post('/info', async (_, res) => {
+    logger.debug('Call to Kiali information');
+    const response = await kialiAPI.fetchInfo();
+    res.json(response);
+  });
+
+  router.post('/config', async (_, res) => {
     logger.debug('Call to Configuration');
     const response = await kialiAPI.fetchConfig();
     res.json(response);
   });
 
-  router.get('/overview', async (req, res) => {
+  router.post('/overview', async (req, res) => {
+    const entity = await getEntityByReq(req);
     const query: OverviewQuery = {
-      annotation: getAnnotations(req.query),
-      duration: Number(req.query.duration) || 600,
-      direction: (req.query.direction as string) || 'inbound',
-      overviewType: (req.query.overviewType as string) || 'app',
+      duration: Number(req.body.query.duration) || 600,
+      direction: (req.body.query.direction as string) || 'inbound',
+      overviewType: (req.body.query.overviewType as string) || 'app',
     };
     logger.debug('Call to Overview');
-    const response = await kialiAPI.fetchOverviewNamespaces(query);
+    const response = await kialiAPI.fetchOverviewNamespaces(entity, query);
     res.json(response);
   });
 
@@ -72,6 +106,7 @@ export async function createRouter(
 ): Promise<express.Router> {
   const { logger } = options;
   const { config } = options;
+  const { catalogApi } = options;
 
   logger.info('Initializing Kiali backend');
 
@@ -79,5 +114,5 @@ export async function createRouter(
 
   const kialiAPI = new KialiApiImpl({ logger, kiali });
 
-  return makeRouter(logger, kialiAPI);
+  return makeRouter(logger, kialiAPI, catalogApi);
 }