+++ plugins/kiali-backend/src/service/standaloneServer.ts
@@ -1,4 +1,5 @@
-import { createServiceBuilder } from '@backstage/backend-common';
+import { createServiceBuilder, HostDiscovery } from '@backstage/backend-common';
+import { CatalogClient } from '@backstage/catalog-client';
 import { ConfigReader } from '@backstage/config';
 
 import { Logger } from 'winston';
@@ -17,10 +18,15 @@ export async function startStandaloneServer(
   options: ServerOptions,
 ): Promise<Server> {
   const logger = options.logger.child({ service: 'kiali-backend' });
+  const config = new ConfigReader({});
+  const catalogApi = new CatalogClient({
+    discoveryApi: HostDiscovery.fromConfig(config),
+  });
   logger.debug('Starting application server...');
   const router = await createRouter({
     logger,
-    config: new ConfigReader({}),
+    config,
+    catalogApi,
   });
 
   let service = createServiceBuilder(module)